requirements.txt
pip==23.2.1
python-dotenv
reflex==0.6.5
pytz==2023.3
supabase